+++ app/src/main/java/com/cs446/housematehub/userinit/SignUpActivity.java
@@ -1,4 +1,4 @@
-package com.cs446.housematehub;
+package com.cs446.housematehub.userinit;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -11,6 +11,8 @@ import android.widget.Toast;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.appcompat.widget.Toolbar;
 
+import com.cs446.housematehub.R;
+import com.cs446.housematehub.houseinit.LoginHouseActivity;
 import com.mobsandgeeks.saripaar.ValidationError;
 import com.mobsandgeeks.saripaar.Validator;
 import com.mobsandgeeks.saripaar.annotation.ConfirmPassword;
